[FEAT #60 ] 로그인 시 쿠키에 role 추가, JwtProvider -> CookieUtil Refactor
