--- gateway-service/src/main/java/com/farmted/gatewayservice/config/AuthorizationFilter.java
@@ -1,86 +1,0 @@
-package com.farmted.gatewayservice.config;
-
-import io.jsonwebtoken.JwtParser;
-import io.jsonwebtoken.Jwts;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.http.HttpHeaders;
-import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.core.env.Environment;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.http.server.reactive.ServerHttpResponse;
-import org.springframework.stereotype.Component;
-import org.springframework.web.server.ServerWebExchange;
-import reactor.core.publisher.Mono;
-
-
-@Component
-@Slf4j
-public class AuthorizationFilter extends AbstractGatewayFilterFactory<AuthorizationFilter.Config> {
-
-    Environment env;
-
-    public AuthorizationFilter(Environment env) {
-        super(Config.class);
-        this.env = env;
-    }
-
-    public static class Config {
-
-    }
-
-    @Override
-    public GatewayFilter apply(Config config) {
-        return ((exchange, chain) -> {
-            ServerHttpRequest request = exchange.getRequest();
-
-
-            if (!request.getHeaders().containsKey(HttpHeaders.AUTHORIZATION)) {
-                return onError(exchange, "No authorization header", HttpStatus.UNAUTHORIZED);
-            }
-
-            String authorizationHeader = request.getHeaders().get(HttpHeaders.AUTHORIZATION).get(0);
-            String jwt = authorizationHeader.replace("Authorization=Bearer-", "");
-
-            if (!isJwtValid(jwt)) {
-                return onError(exchange, "JWT token is not valid", HttpStatus.UNAUTHORIZED);
-            }
-
-            return chain.filter(exchange);
-        });
-    }
-
-    // Mono, flux -> Spring WebFlux
-    // 클라이언트 요청이 들어왔을 때 반환 시켜주는 데이터 타입, 단일값일 경우 Mono, 다중값일 경우 flux
-    private Mono<Void> onError(ServerWebExchange exchange, String err, HttpStatus httpStatus) {
-        ServerHttpResponse response = exchange.getResponse();
-        response.setStatusCode(httpStatus);
-
-        log.error(err);
-        return response.setComplete();
-    }
-
-    private boolean isJwtValid(String jwt) {
-        boolean returnValue = true;
-
-        String subject = null;
-
-        try {
-        JwtParser jwtParser = Jwts.parserBuilder()
-                .setSigningKey(env.getProperty("jwt-secret-key"))
-                .build();
-
-        subject = jwtParser.parseClaimsJws(jwt).getBody().getSubject();
-
-        } catch (Exception ex) {
-            returnValue = false;
-        }
-
-        if (subject == null || subject.isEmpty()) {
-            returnValue = false;
-        }
-
-        return returnValue;
-    }
-}
