+++ pass-service/src/main/java/com/farmted/passservice/service/PassService.java
@@ -19,4 +19,5 @@ public interface PassService {
     String findUuidByEmail(String email);
     String reIssue(String uuid);
     void changeRoleByMemberService(String uuid);
+    String setRole(RequestLoginDto dto);
 }
