+++ pass-service/src/main/java/com/farmted/passservice/util/oauth2/handler/OAuth2LoginSuccessHandler.java
@@ -2,6 +2,7 @@ package com.farmted.passservice.util.oauth2.handler;
 
 import com.farmted.passservice.enums.RoleEnums;
 import com.farmted.passservice.enums.TokenType;
+import com.farmted.passservice.util.cookies.CookieUtil;
 import com.farmted.passservice.util.jwt.JwtProvider;
 import com.farmted.passservice.util.oauth2.CustomOAuth2User;
 import jakarta.servlet.ServletException;
@@ -21,6 +22,7 @@ import java.io.IOException;
 public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
 
     private final JwtProvider jwtProvider;
+    private final CookieUtil cookieUtil;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
@@ -33,12 +35,12 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
 
             // Pass의 Role이 GUEST일 경우 회원가입이 완료되지 않은 회원이므로 회원 상세정보 적기 페이지로 리다이렉트
             if (role == RoleEnums.GUEST) {
-                jwtProvider.setToken(token, TokenType.ACCESS, response);
+                cookieUtil.setToken(token, TokenType.ACCESS, response);
                 String refreshToken = jwtProvider.createToken(uuid, role, TokenType.REFRESH);
                 jwtProvider.saveRefreshToken(uuid, refreshToken);
                 response.sendRedirect("/home");
             } else {
-                jwtProvider.setToken(token, TokenType.ACCESS, response);
+                cookieUtil.setToken(token, TokenType.ACCESS, response);
                 jwtProvider.saveRefreshToken(uuid, token);
                 response.sendRedirect("/home");
             }
